+++ core/arch/arm/mm/core_mmu_lpae.c
@@ -77,8 +77,6 @@
 #include <types_ext.h>
 #include <util.h>
 
-#include "core_mmu_private.h"
-
 #ifndef DEBUG_XLAT_TABLE
 #define DEBUG_XLAT_TABLE 0
 #endif
